added AuthService to secure endpoints that allowed for admins only

+++ quizz-web/src/main/java/us/quizz/endpoints/QuizEndpoint.java
@@ -1,21 +1,23 @@
 package us.quizz.endpoints;
 
-import com.google.api.server.spi.config.Api;
-import com.google.api.server.spi.config.ApiMethod;
-import com.google.api.server.spi.config.ApiMethod.HttpMethod;
-import com.google.api.server.spi.response.CollectionResponse;
-import com.google.inject.Inject;
-
-import us.quizz.entities.Question;
-import us.quizz.entities.Quiz;
-import us.quizz.service.QuestionService;
-import us.quizz.service.QuizService;
-
 import java.util.Map;
 import java.util.Set;
 
 import javax.annotation.Nullable;
 import javax.inject.Named;
+
+import us.quizz.entities.Question;
+import us.quizz.entities.Quiz;
+import us.quizz.service.AuthService;
+import us.quizz.service.QuestionService;
+import us.quizz.service.QuizService;
+
+import com.google.api.server.spi.config.Api;
+import com.google.api.server.spi.config.ApiMethod;
+import com.google.api.server.spi.config.ApiMethod.HttpMethod;
+import com.google.api.server.spi.response.CollectionResponse;
+import com.google.api.server.spi.response.ForbiddenException;
+import com.google.inject.Inject;
 
 @Api(name = "quizz", description = "The API for Quizz.us", version = "v1")
 public class QuizEndpoint {
@@ -23,11 +25,13 @@ public class QuizEndpoint {
 
   private QuizService quizService;
   private QuestionService questionService;
+  private AuthService authService;
 
   @Inject
-  public QuizEndpoint(QuizService quizService, QuestionService questionService) {
+  public QuizEndpoint(QuizService quizService, QuestionService questionService, AuthService authService) {
     this.quizService = quizService;
     this.questionService = questionService;
+    this.authService = authService;
   }
 
   // Lists the quiz in datastore using paging.
@@ -42,7 +46,11 @@ public class QuizEndpoint {
 
   // Sets the quiz for the quizID given to be shown on landing page by default.
   @ApiMethod(name = "showQuiz", path = "showQuiz", httpMethod = HttpMethod.GET)
-  public Quiz showQuiz(@Named("quizID") String quizID) {
+  public Quiz showQuiz(@Named("quizID") String quizID) throws ForbiddenException {
+    if (!authService.isUserAdmin()) {
+      throw new ForbiddenException("Forbidden");
+    }
+
     Quiz quiz = quizService.get(quizID);
     quiz.setShowOnDefault(true);
     return quizService.save(quiz);
@@ -50,7 +58,11 @@ public class QuizEndpoint {
 
   // Sets the quiz for the quizID given to be hidden on landing page by default.
   @ApiMethod(name = "hideQuiz", path = "hideQuiz", httpMethod = HttpMethod.GET)
-  public Quiz hideQuiz(@Named("quizID") String quizID) {
+  public Quiz hideQuiz(@Named("quizID") String quizID) throws ForbiddenException {
+    if (!authService.isUserAdmin()) {
+      throw new ForbiddenException("Forbidden");
+    }
+
     Quiz quiz = quizService.get(quizID);
     quiz.setShowOnDefault(false);
     return quizService.save(quiz);
@@ -73,7 +85,11 @@ public class QuizEndpoint {
   // Inserts a new entity into Datastore. If the entity already exists, an exception will be thrown.
   // If the showOnDefault field is not filled, it will be set to false.
   @ApiMethod(name = "insertQuiz", path = "insertQuiz", httpMethod = HttpMethod.POST)
-  public Quiz insertQuiz(Quiz quiz) {
+  public Quiz insertQuiz(Quiz quiz) throws ForbiddenException {
+    if (!authService.isUserAdmin()) {
+      throw new ForbiddenException("Forbidden");
+    }
+
     if (quiz.getShowOnDefault() == null) {
       quiz.setShowOnDefault(false);
     }
@@ -84,7 +100,12 @@ public class QuizEndpoint {
   // such as Question, Answer, and UserAnswer.
   @ApiMethod(name = "removeQuizRecursively", path = "removeQuizRecursively",
              httpMethod = HttpMethod.DELETE)
-  public void removeQuizRecursively(@Named("quizID") String quizID) {
+  public void removeQuizRecursively(@Named("quizID") String quizID) throws ForbiddenException {
+    if (!authService.isUserAdmin()) {
+      throw new ForbiddenException("Forbidden");
+    }
+
     quizService.deleteRecursively(quizID);
   }
+
 }
